+++ elmyra.ip.access.epo/CHANGES.rst
@@ -11,6 +11,7 @@ development
 - ui: fix project name display in liveview mode
 - ui: enhance basket activation/deactivation in case project has no basket
 - middleware: use static signing key for opaque parameter subsystem
+- ui, middleware: show expiry time and improve statusline formatting in liveview mode
 
 0.24.3
 ======